stm32: move ptp clock under v2
